Some changes here.
As observed, gcc 6.3 has weird behaviour when using sanitiser flags, and triggered a stack overflow on an internal malloc in vasprintf when milliseconds timing was added to logging. This error was bypassed by adding a simple printf in front if it.
On top of that, some functions errored on asan when calling logging macros. This behaviour was mainly observed in small functions, so I extracted error interpretation and logging to the calling function.

Changes :
1. Extracted some LOG calls to calling functions.
2. reimplemented log_s_vasprintf and renamed to log_s_vsnprintf. Since we already have a buffer to print in, we don't need an intermediary allocated buffer. Using vsnprintf() from now on.
3. log_buffs.log_build has to initialised or reading its value would be interpreted with random memory
4. Renamed log_s_vasprintf to log_vsnprintf
5. Added code to use millisecond precision on logs. Not used for now since it triggers weird asan errors elsewhere in the code.